More shuffling around to avoid importing extension code into tests
