+++ src/extension/commands/sdk.ts
@@ -4,9 +4,12 @@ import * as os from "os";
 import * as path from "path";
 import * as vs from "vscode";
 import { ProgressLocation, Uri, window } from "vscode";
+import { flutterPath, pubPath } from "../../shared/constants";
 import { LogCategory, LogSeverity } from "../../shared/enums";
 import { Sdks } from "../../shared/interfaces";
 import { flatMap, PromiseCompleter } from "../../shared/utils";
+import { sortBy } from "../../shared/utils/array";
+import { getChildFolders, hasPubspec } from "../../shared/utils/fs";
 import { fsPath } from "../../shared/vscode/utils";
 import { WorkspaceContext } from "../../shared/workspace";
 import { config } from "../config";
@@ -21,10 +24,8 @@ import { Stagehand, StagehandTemplate } from "../pub/stagehand";
 import { FlutterSampleSnippet, getFlutterSnippets } from "../sdk/flutter_docs_snippets";
 import { createFlutterSampleInTempFolder } from "../sdk/flutter_samples";
 import { DartSdkManager, FlutterSdkManager } from "../sdk/sdk_manager";
-import { flutterPath, pubPath, showDartActivationFailure, showFlutterActivationFailure } from "../sdk/utils";
+import { showDartActivationFailure, showFlutterActivationFailure } from "../sdk/utils";
 import * as util from "../utils";
-import { sortBy } from "../utils/array";
-import { getChildFolders, hasPubspec } from "../utils/fs";
 import { log, logError, logProcess } from "../utils/log";
 import { globalFlutterArgs, safeSpawn } from "../utils/processes";
 import * as channels from "./channels";
