+++ src/extension/debug/dart_debug_impl.ts
@@ -6,8 +6,8 @@ import { DebugProtocol } from "vscode-debugprotocol";
 import { pleaseReportBug } from "../../shared/constants";
 import { LogCategory, LogSeverity } from "../../shared/enums";
 import { flatMap, throttle, uniq, uriToFilePath } from "../../shared/utils";
+import { white } from "../../shared/utils/colors";
 import { config } from "../config";
-import { white } from "../utils/colors";
 import { getLogHeader, logError, logWarn } from "../utils/log";
 import { safeSpawn } from "../utils/processes";
 import { DebuggerResult, ObservatoryConnection, SourceReportKind, VM, VMClass, VMClassRef, VMErrorRef, VMEvent, VMFrame, VMInstance, VMInstanceRef, VMIsolate, VMIsolateRef, VMLibrary, VMMapEntry, VMObj, VMScript, VMScriptRef, VMSentinel, VMSourceReport, VMStack, VMTypeRef } from "./dart_debug_protocol";
