+++ src/extension/debug/threads.ts
@@ -2,7 +2,7 @@ import { Thread, ThreadEvent } from "vscode-debugadapter";
 import { DebugProtocol } from "vscode-debugprotocol";
 import { LogCategory } from "../../shared/enums";
 import { PromiseCompleter } from "../../shared/utils";
-import { isKnownInfrastructureThread } from "../utils/debugger";
+import { isKnownInfrastructureThread } from "../../shared/utils/debugger";
 import { logError } from "../utils/log";
 import { DartDebugSession, InstanceWithEvaluateName, VmExceptionMode } from "./dart_debug_impl";
 import { DebuggerResult, VMBreakpoint, VMInstanceRef, VMIsolate, VMIsolateRef, VMResponse, VMScript, VMScriptRef } from "./dart_debug_protocol";
