+++ src/extension/extension.ts
@@ -2,11 +2,11 @@ import * as fs from "fs";
 import * as path from "path";
 import { isArray } from "util";
 import * as vs from "vscode";
-import { flutterExtensionIdentifier } from "../shared/constants";
+import { analyzerSnapshotPath, dartPlatformName, dartVMPath, flutterExtensionIdentifier, flutterPath, isWin, platformDisplayName } from "../shared/constants";
 import { LogCategory } from "../shared/enums";
 import { Sdks } from "../shared/interfaces";
 import { internalApiSymbol } from "../shared/symbols";
-import { dartPlatformName, forceWindowsDriveLetterToUppercase, isWin, isWithinPath, platformDisplayName } from "../shared/utils";
+import { forceWindowsDriveLetterToUppercase, isWithinPath } from "../shared/utils";
 import { InternalExtensionApi } from "../shared/vscode/interfaces";
 import { fsPath } from "../shared/vscode/utils";
 import { Context } from "../shared/vscode/workspace";
@@ -66,7 +66,7 @@ import { isPubGetProbablyRequired, promptToRunPubGet } from "./pub/pub";
 import { DartCapabilities } from "./sdk/capabilities";
 import { StatusBarVersionTracker } from "./sdk/status_bar_version_tracker";
 import { checkForStandardDartSdkUpdates } from "./sdk/update_check";
-import { analyzerSnapshotPath, dartVMPath, flutterPath, handleMissingSdks, initWorkspace } from "./sdk/utils";
+import { handleMissingSdks, initWorkspace } from "./sdk/utils";
 import { DartUriHandler } from "./uri_handlers/uri_handler";
 import { showUserPrompts } from "./user_prompts";
 import * as util from "./utils";
