+++ src/extension/providers/debug_config_provider.ts
@@ -4,9 +4,9 @@ import * as path from "path";
 import * as vs from "vscode";
 import { CancellationToken, DebugConfiguration, DebugConfigurationProvider, ProviderResult, Uri, window, workspace, WorkspaceFolder } from "vscode";
 import { DebugSession } from "vscode-debugadapter";
-import { CHROME_OS_VM_SERVICE_PORT } from "../../shared/constants";
+import { CHROME_OS_VM_SERVICE_PORT, dartVMPath, flutterPath, isChromeOS, pubPath, pubSnapshotPath } from "../../shared/constants";
 import { Sdks } from "../../shared/interfaces";
-import { forceWindowsDriveLetterToUppercase, isChromeOS, isWithinPath } from "../../shared/utils";
+import { forceWindowsDriveLetterToUppercase, isWithinPath } from "../../shared/utils";
 import { fsPath } from "../../shared/vscode/utils";
 import { Analytics } from "../analytics";
 import { LastDebugSession } from "../commands/debug";
@@ -25,7 +25,6 @@ import { Device } from "../flutter/flutter_types";
 import { locateBestProjectRoot } from "../project";
 import { PubGlobal } from "../pub/global";
 import { WebDev } from "../pub/webdev";
-import { dartVMPath, flutterPath, pubPath, pubSnapshotPath } from "../sdk/utils";
 import { checkProjectSupportsPubRunTest, isDartFile, isFlutterProjectFolder, isFlutterWebProjectFolder, isFlutterWorkspaceFolder, isInsideFolderNamed, isTestFile, isTestFileOrFolder } from "../utils";
 import { log, logError, logWarn } from "../utils/log";
 import { TestResultsProvider } from "../views/test_view";
