+++ src/extension/sdk/dev_tools.ts
@@ -3,10 +3,10 @@ import * as os from "os";
 import * as path from "path";
 import * as vs from "vscode";
 import { window, workspace } from "vscode";
-import { CHROME_OS_DEVTOOLS_PORT, pleaseReportBug } from "../../shared/constants";
+import { CHROME_OS_DEVTOOLS_PORT, isChromeOS, pleaseReportBug, pubPath } from "../../shared/constants";
 import { FlutterService, LogCategory } from "../../shared/enums";
 import { Sdks } from "../../shared/interfaces";
-import { isChromeOS } from "../../shared/utils";
+import { waitFor } from "../../shared/utils/promises";
 import { Analytics } from "../analytics";
 import { DebugCommands, debugSessions } from "../commands/debug";
 import { config } from "../config";
@@ -14,9 +14,7 @@ import { PubGlobal } from "../pub/global";
 import { StdIOService, UnknownNotification } from "../services/stdio_service";
 import { getRandomInt } from "../utils";
 import { log, logError } from "../utils/log";
-import { waitFor } from "../utils/promises";
 import { DartDebugSessionInformation } from "../utils/vscode/debug";
-import { pubPath } from "./utils";
 
 const devtools = "devtools";
 const devtoolsPackageName = "Dart DevTools";
