+++ src/extension/sdk/sdk_manager.ts
@@ -1,11 +1,11 @@
 import * as fs from "fs";
 import * as path from "path";
 import * as vs from "vscode";
+import { dartVMPath, flutterPath } from "../../shared/constants";
 import { Sdks } from "../../shared/interfaces";
+import { getChildFolders } from "../../shared/utils/fs";
 import { config } from "../config";
 import { getSdkVersion, versionIsAtLeast } from "../utils";
-import { getChildFolders } from "../utils/fs";
-import { dartVMPath, flutterPath } from "./utils";
 
 abstract class SdkManager {
 	protected sdks: Sdks;
