+++ src/extension/sdk/update_check.ts
@@ -1,9 +1,9 @@
 import { window } from "vscode";
+import { DART_DOWNLOAD_URL } from "../../shared/constants";
 import { WorkspaceContext } from "../../shared/workspace";
 import { config } from "../config";
 import { getLatestSdkVersion, openInBrowser, versionIsAtLeast } from "../utils";
 import { logError } from "../utils/log";
-import { DART_DOWNLOAD_URL } from "./utils";
 
 export async function checkForStandardDartSdkUpdates(workspaceContext: WorkspaceContext): Promise<void> {
 	if (!config.checkForSdkUpdates || !workspaceContext.hasOnlyDartProjects)
