+++ src/extension/user_prompts.ts
@@ -3,9 +3,8 @@ import * as os from "os";
 import * as path from "path";
 import { Context } from "vm";
 import * as vs from "vscode";
-import { doNotAskAgainAction, flutterExtensionIdentifier, flutterSurvey2019Q2PromptWithAnalytics, flutterSurvey2019Q2PromptWithoutAnalytics, longRepeatPromptThreshold, noRepeatPromptThreshold, noThanksAction, openDevToolsAction, takeSurveyAction, wantToTryDevToolsPrompt } from "../shared/constants";
+import { doNotAskAgainAction, flutterExtensionIdentifier, flutterSurvey2019Q2PromptWithAnalytics, flutterSurvey2019Q2PromptWithoutAnalytics, isWin, longRepeatPromptThreshold, noRepeatPromptThreshold, noThanksAction, openDevToolsAction, takeSurveyAction, wantToTryDevToolsPrompt } from "../shared/constants";
 import { LogCategory, LogSeverity } from "../shared/enums";
-import { isWin } from "../shared/utils";
 import { fsPath } from "../shared/vscode/utils";
 import { WorkspaceContext } from "../shared/workspace";
 import { markProjectCreationEnded, markProjectCreationStarted } from "./commands/sdk";
