+++ src/extension/utils/log.ts
@@ -2,8 +2,8 @@ import * as child_process from "child_process";
 import * as fs from "fs";
 import * as os from "os";
 import * as path from "path";
+import { platformEol } from "../../shared/constants";
 import { LogCategory, LogSeverity } from "../../shared/enums";
-import { platformEol } from "../../shared/utils";
 import { config } from "../config";
 import { IAmDisposable, LogEmitter, LogMessage } from "../debug/utils";
 import { getRandomInt } from "../utils";
