+++ src/shared/utils/array.ts
