+++ src/test/dart_debug_client.ts
@@ -3,11 +3,11 @@ import { SpawnOptions } from "child_process";
 import { DebugSession, DebugSessionCustomEvent } from "vscode";
 import { DebugProtocol } from "vscode-debugprotocol";
 import { debugSessions } from "../extension/commands/debug";
-import { not } from "../extension/utils/array";
-import { isKnownInfrastructureThread } from "../extension/utils/debugger";
 import { log } from "../extension/utils/log";
 import { DebugCommandHandler } from "../shared/interfaces";
 import { Notification, Test, TestDoneNotification, TestStartNotification } from "../shared/test_protocol";
+import { not } from "../shared/utils/array";
+import { isKnownInfrastructureThread } from "../shared/utils/debugger";
 import { TestResultsProvider } from "../shared/vscode/interfaces";
 import { DebugClient, ILocation, IPartialLocation } from "./debug_client_ms";
 import { delay, watchPromise, withTimeout } from "./helpers";
