+++ src/test/dart_only/debug/dart_cli.test.ts
@@ -5,8 +5,8 @@ import { config } from "../../../extension/config";
 import { debugAnywayAction, showErrorsAction } from "../../../extension/providers/debug_config_provider";
 import { getRandomInt } from "../../../extension/utils";
 import { log } from "../../../extension/utils/log";
+import { platformEol } from "../../../shared/constants";
 import { fetch } from "../../../shared/fetch";
-import { platformEol } from "../../../shared/utils";
 import { fsPath } from "../../../shared/vscode/utils";
 import { DartDebugClient } from "../../dart_debug_client";
 import { ensureFrameCategories, ensureMapEntry, ensureVariable, ensureVariableWithIndex, isExternalPackage, isLocalPackage, isSdkFrame, isUserCode, spawnDartProcessPaused } from "../../debug_helpers";
