+++ src/test/dart_only/debug/dart_test.test.ts
@@ -4,9 +4,9 @@ import * as vs from "vscode";
 import { DebugProtocol } from "vscode-debugprotocol";
 import { OpenFileTracker } from "../../../extension/analysis/open_file_tracker";
 import { log, logInfo } from "../../../extension/utils/log";
-import { makeRegexForTest } from "../../../extension/utils/test";
 import { TestOutlineVisitor } from "../../../extension/utils/vscode/outline";
 import { TestStatus } from "../../../shared/enums";
+import { makeRegexForTest } from "../../../shared/utils/test";
 import { fsPath } from "../../../shared/vscode/utils";
 import { DartDebugClient } from "../../dart_debug_client";
 import { activate, defer, delay, ext, extApi, getExpectedResults, getLaunchConfiguration, getPackages, helloWorldTestBrokenFile, helloWorldTestDupeNameFile, helloWorldTestMainFile, helloWorldTestSkipFile, helloWorldTestTreeFile, makeTextTree, openFile, positionOf, withTimeout } from "../../helpers";
