+++ src/test/helpers.ts
@@ -7,15 +7,15 @@ import * as sinon from "sinon";
 import * as vs from "vscode";
 import { DelayedCompletionItem } from "../extension/providers/dart_completion_item_provider";
 import { isAnalyzable, vsCodeVersionConstraint } from "../extension/utils";
-import { tryDeleteFile } from "../extension/utils/fs";
 import { log, logError, logTo, logWarn } from "../extension/utils/log";
-import { waitFor } from "../extension/utils/promises";
 import { PackageDep } from "../extension/views/packages_view";
 import { SuiteTreeItem } from "../extension/views/test_view";
 import { dartCodeExtensionIdentifier } from "../shared/constants";
 import { LogCategory, LogSeverity, TestStatus } from "../shared/enums";
 import { internalApiSymbol } from "../shared/symbols";
 import { flatMap } from "../shared/utils";
+import { tryDeleteFile } from "../shared/utils/fs";
+import { waitFor } from "../shared/utils/promises";
 import { InternalExtensionApi, TestItemTreeItem, TestResultsProvider } from "../shared/vscode/interfaces";
 import { fsPath } from "../shared/vscode/utils";
 import { Context } from "../shared/vscode/workspace";
